Enable several error prone checks

+++ WEB-INF/classes/com/cohort/array/FloatArray.java
@@ -422,7 +422,7 @@ public class FloatArray extends PrimitiveArray {
    */
   @Override
   public void addNLongs(final int n, final long value) {
-    addN(n, value); // ! assumes value=Integer.MAX_VALUE isn't maxIsMV
+    addN(n, (float) value); // ! assumes value=Integer.MAX_VALUE isn't maxIsMV
   }
 
   /**
@@ -730,7 +730,7 @@ public class FloatArray extends PrimitiveArray {
    */
   @Override
   public void setLong(final int index, final long i) {
-    set(index, i);
+    set(index, (float) i);
   }
 
   /**
